requirements.txt
openai-whisper==20231117
ffmpeg-python==0.2.0
pydub==0.25.1
numpy==1.26.4